requirements.txt
lxml==5.2.1
pillow==10.3.0
python-docx==1.1.0
typing_extensions==4.10.0
